+++ src/components/Room/MarkerIcon/MarkerIcon.module.scss
