+++ src/components/Room/RoomList/RoomList.module.scss
@@ -1,0 +1,85 @@
+@use "src/utils/scss/media" as media;
+
+.list {
+  padding-left: 13px;
+  overflow: scroll;
+  height: 878px;
+  display: grid;
+  grid-template-columns: repeat(2, 173px);
+  gap: 15px;
+  -ms-overflow-style: none;
+  scrollbar-width: none;
+
+  &::-webkit-scrollbar {
+    display: none;
+  }
+
+  @include media.media-breakpoint(mobile) {
+    margin-top: 10px;
+    grid-template-columns: repeat(2, 1fr);
+    padding: 0 13px;
+  }
+
+  &__box {
+    box-shadow: 0 3px 6px 0 rgba(0 0 0 / 16%);
+    border: 1px #d2dae2 solid;
+    cursor: pointer;
+
+    &:hover {
+      border: 1px solid #f7941e;
+    }
+  }
+
+  &__name {
+    height: 96px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    font-size: 16px;
+    font-weight: bold;
+    color: #252525;
+
+    @include media.media-breakpoint(mobile) {
+      height: 61px;
+    }
+  }
+}
+
+.fee {
+  background-color: #f9f9f9;
+  font-size: 12px;
+  color: #858585;
+  padding: 10px;
+
+  &__info {
+    display: grid;
+    grid-template-columns: 50px 1fr;
+    align-items: center;
+
+    span:nth-child(2) {
+      overflow: hidden;
+      text-overflow: ellipsis;
+      white-space: nowrap;
+    }
+  }
+
+  &__monthly {
+    background-color: #4db297;
+    width: 40px;
+    text-align: center;
+    border-radius: 15px;
+    line-height: 2;
+    color: #f3f3f3;
+  }
+
+  &__charter {
+    margin-top: 4px;
+    background-color: #f7941e;
+    width: 40px;
+    text-align: center;
+    margin-right: 9px;
+    border-radius: 15px;
+    line-height: 2;
+    color: #f3f3f3;
+  }
+}
