+++ src/pages/Room/RoomPage/index.tsx
@@ -3,9 +3,10 @@ import {
 } from 'react-naver-maps';
 import ReactDOMServer from 'react-dom/server';
 import useMediaQuery from 'utils/hooks/useMediaQuery';
+import MarkerIcon from 'components/Room/MarkerIcon';
+import RoomList from 'components/Room/RoomList';
 import useRoomList from './hooks/useRoomList';
 import styles from './RoomPage.module.scss';
-import MarkerIcon from './component/MarkerIcon';
 
 function RoomPage() {
   const isMobile = useMediaQuery();
@@ -41,23 +42,7 @@ function RoomPage() {
             ))}
           </NaverMap>
         </MapDiv>
-        <div className={styles.list}>
-          {roomList?.lands.map((room) => (
-            <div className={styles.list__box} key={room.id}>
-              <div className={styles.list__name}>{room.name}</div>
-              <div className={styles.fee}>
-                <div className={styles.fee__info}>
-                  <span className={styles.fee__monthly}>월세</span>
-                  <span>{room.monthly_fee ? room.monthly_fee : '정보없음'}</span>
-                </div>
-                <div className={styles.fee__info}>
-                  <span className={styles.fee__charter}>전세</span>
-                  <span>{room.charter_fee ? `${room.charter_fee}만원` : ' - '}</span>
-                </div>
-              </div>
-            </div>
-          ))}
-        </div>
+        <RoomList lands={roomList?.lands} />
       </div>
     </div>
   );
